requirements.txt
ultralytics>=8.0.0
roboflow>=1.1.9
python-dotenv>=1.0.1
rich>=13.7.1
typer>=0.12.5
opencv-python-headless>=4.9.0.80
numpy>=1.24.0
PyYAML>=6.0.1
